requirements.txt
# Soccer Data Scraping
# Nota: Python 3.14 es muy nuevo, usamos versión 1.5.1 que es compatible
soccerdata==1.5.1

# Telegram Bot
python-telegram-bot>=20.0

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Scheduling
schedule>=1.2.0
python-dateutil>=2.8.0
pytz>=2023.3

# Environment Variables
python-dotenv>=1.0.0

# Database
# sqlite3 viene incluido con Python

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Logging
colorlog>=6.7.0
